Harden Lieferzeiten permission boundaries and admin action guards

+++ custom/plugins/LieferzeitenAdmin/src/Resources/app/administration/src/module/lieferzeiten-settings/page/lieferzeiten-notification-toggle-list/index.js
@@ -32,6 +32,10 @@ Component.register('lieferzeiten-notification-toggle-list', {
     },
 
     computed: {
+        hasEditAccess() {
+            return this.acl.can('lieferzeiten.editor') || this.acl.can('admin');
+        },
+
         columns() {
             return [
                 { property: 'triggerKey', label: 'Trigger', inlineEdit: 'string', primary: true },
@@ -69,6 +73,10 @@ Component.register('lieferzeiten-notification-toggle-list', {
             this.getList();
         },
         onInlineEditSave(item) {
+            if (!this.hasEditAccess) {
+                return Promise.resolve();
+            }
+
             this.isLoading = true;
             item.code = `${item.triggerKey}:${item.channel}`;
             return this.repository.save(item, Shopware.Context.api).then(() => {
@@ -76,12 +84,20 @@ Component.register('lieferzeiten-notification-toggle-list', {
             });
         },
         onDelete(item) {
+            if (!this.hasEditAccess) {
+                return Promise.resolve();
+            }
+
             this.isLoading = true;
             return this.repository.delete(item.id, Shopware.Context.api).then(() => {
                 this.getList();
             });
         },
         onCreate() {
+            if (!this.hasEditAccess) {
+                return Promise.resolve();
+            }
+
             const entity = this.repository.create(Shopware.Context.api);
             entity.triggerKey = 'commande.creee';
             entity.channel = 'email';
